[fix] remove duplicate fixture. fix tests
